// apps/lab-app/src/types/doctor/doctor.ts
export interface Doctor {
    id?: number;
    name: string;
    email?: string;
    speciality?: string;
    qualification?: string;
    hospitalAffiliation?: string;
    licenseNumber?: string;
    phone?: number;
    address?: string;
    city?: string;
    state?: string;
    country?: string;
}